+++ app/mailers/analysis_mailer.rb
@@ -1,14 +1,14 @@
 # coding: utf-8
+
 class AnalysisMailer < ApplicationMailer
-  default :from => 'Leonis.0813@gmail.com'
+  default from: 'Leonis.0813@gmail.com'
 
-  def finished(analysis, is_success)
+  def completed(analysis)
     @analysis = analysis
-    subject = is_success ? '分析が完了しました' : '分析中にエラーが発生しました'
-    template_name = is_success ? 'success' : 'failer'
-    tmp_dir = File.join(Rails.root, "tmp/models/#{analysis.id}")
+    subject = '分析が完了しました'
+    tmp_dir = Rails.root.join('tmp', 'models', analysis.id.to_s)
 
-    Dir.mktmpdir(nil, File.join(Rails.root, 'tmp/files')) do |dir|
+    Dir.mktmpdir(nil, Rails.root.join('tmp', 'files')) do |dir|
       zip_file_name = File.join(dir, 'analysis.zip')
 
       Zip::File.open(zip_file_name, Zip::File::CREATE) do |zip|
@@ -19,7 +19,13 @@ class AnalysisMailer < ApplicationMailer
       end
 
       attachments['analysis.zip'] = File.read(zip_file_name)
-      mail(:to => 'Leonis.0813@gmail.com', :subject => subject, :template_name => template_name)
+      mail(to: 'Leonis.0813@gmail.com', subject: subject, template_name: 'success')
     end
+  end
+
+  def error(analysis)
+    @analysis = analysis
+    subject = '分析中にエラーが発生しました'
+    mail(to: 'Leonis.0813@gmail.com', subject: subject, template_name: 'failer')
   end
 end