+++ app/models/analysis.rb
@@ -1,7 +1,17 @@
 class Analysis < ActiveRecord::Base
+  DEFAULT_PAIR = 'USDJPY'.freeze
+  PAIRS = %w[AUDJPY CADJPY CHFJPY EURJPY EURUSD GBPJPY NZDJPY USDJPY].freeze
+  STATES = %w[processing completed error].freeze
+
   validate :valid_period?
-  validates :batch_size, :numericality => {:only_integer => true, :greater_than => 0, :message => 'invalid'}
-  validates :state, :inclusion => {:in => %w[ processing completed ], :message => 'invalid'}
+  validates :batch_size, :pair, :state,
+            presence: {message: 'absent'}
+  validates :batch_size,
+            numericality: {only_integer: true, greater_than: 0, message: 'invalid'}
+  validates :pair,
+            inclusion: {in: PAIRS, message: 'invalid'}
+  validates :state,
+            inclusion: {in: STATES, message: 'invalid'}
 
   private
 
@@ -10,10 +20,9 @@ class Analysis < ActiveRecord::Base
     errors.add(:to, 'invalid') unless to
 
     return if errors.messages.include?(:from) or errors.messages.include?(:to)
+    return if from < to
 
-    unless from < to
-      errors.add(:from, 'invalid')
-      errors.add(:to, 'invalid')
-    end
+    errors.add(:from, 'invalid')
+    errors.add(:to, 'invalid')
   end
 end