+++ app/models/prediction.rb
@@ -1,17 +1,27 @@
 class Prediction < ActiveRecord::Base
+  RESULTS = %w[up down range].freeze
+
   validate :valid_period?
-  validates :model, :format => {:with => /\.zip\z/, :message => 'invalid'}
-  validates :result, :inclusion => {:in => %w[ up down range ], :message => 'invalid'}, :allow_nil => true
-  validates :state, :inclusion => {:in => %w[ processing completed ], :message => 'invalid'}
+  validates :model, :state,
+            presence: {message: 'absent'}
+  validates :model,
+            format: {with: /\.zip\z/, message: 'invalid'}
+  validates :pair,
+            inclusion: {in: Analysis::PAIRS, message: 'invalid'},
+            allow_nil: true
+  validates :result,
+            inclusion: {in: RESULTS, message: 'invalid'},
+            allow_nil: true
+  validates :state,
+            inclusion: {in: Analysis::STATES, message: 'invalid'}
 
   private
 
   def valid_period?
     return unless from and to
+    return if from < to
 
-    unless from < to
-      errors.add(:from, 'invalid')
-      errors.add(:to, 'invalid')
-    end
+    errors.add(:from, 'invalid')
+    errors.add(:to, 'invalid')
   end
 end