+++ lib/errors/bad_request.rb
@@ -3,7 +3,7 @@ class BadRequest < StandardError
 
   def initialize(error_codes)
     @errors = Array.wrap(error_codes).map do |error_code|
-      {:error_code => error_code}
+      {error_code: error_code}
     end
   end
 end