+++ spec/rails_helper.rb
@@ -1,9 +1,20 @@
 # This file is copied to spec/ when you run 'rails generate rspec:install'
 require 'spec_helper'
 ENV['RAILS_ENV'] ||= 'test'
-require File.expand_path('../../config/environment', __FILE__)
+ENV['COVERAGE'] ||= 'off'
+
+if ENV['COVERAGE'] == 'on'
+  require 'simplecov'
+  require 'simplecov-rcov'
+  SimpleCov.formatter = SimpleCov::Formatter::RcovFormatter
+  SimpleCov.start 'rails' do
+    SimpleCov.command_name Time.now.utc.to_s
+  end
+end
+
+require File.expand_path('../config/environment', __dir__)
 # Prevent database truncation if the environment is production
-abort("The Rails environment is running in production mode!") if Rails.env.production?
+abort('The Rails environment is running in production mode!') if Rails.env.production?
 require 'rspec/rails'
 # Add additional requires below this line. Rails is not loaded until this point!
 
@@ -20,7 +31,7 @@ require 'rspec/rails'
 # directory. Alternatively, in the individual `*_spec.rb` files, manually
 # require only the support files necessary.
 #
-Dir[Rails.root.join('spec/support/**/*.rb')].each { |f| require f }
+Dir[Rails.root.join('spec', 'support', '**', '*.rb')].each {|f| require f }
 
 # Checks for pending migrations and applies them before tests are run.
 # If you are not using ActiveRecord, you can remove this line.
@@ -58,4 +69,10 @@ RSpec.configure do |config|
   # config.filter_gems_from_backtrace("gem name")
   config.run_all_when_everything_filtered = true
   config.include CommonHelper
+  config.include FactoryBot::Syntax::Methods
+  config.include ViewHelper
+
+  config.before(:suite) do
+    DatabaseCleaner.strategy = :transaction
+  end
 end