+++ spec/support/shareds/common_shared.rb
@@ -1,0 +1,6 @@
+# coding: utf-8
+
+shared_context 'トランザクション作成' do
+  before(:all) { DatabaseCleaner.start }
+  after(:all) { DatabaseCleaner.clean }
+end