+++ spec/views/analyses/manage.html.erb_spec.rb
@@ -1,105 +1,127 @@
 # coding: utf-8
+
 require 'rails_helper'
 
-describe 'analyses/manage', :type => :view do
+describe 'analyses/manage', type: :view do
   per_page = 1
+  default_attribute = {
+    from: 2.months.ago,
+    to: 1.month.ago,
+    pair: Analysis::PAIRS.sample,
+    batch_size: 100,
+    state: 'processing',
+  }
 
-  shared_context '分析ジョブを登録する' do |num|
+  shared_context '分析ジョブを登録する' do |total: per_page, attribute: default_attribute|
     before(:all) do
-      param = {:from => 2.month.ago, :to => 1.month.ago, :batch_size => 100}
-      num.times { Analysis.create!(param.merge(:state => %w[ processing completed ].sample)) }
-      @analyses = Analysis.order(:created_at => :desc).page(1)
+      total.times { Analysis.create!(attribute) }
+      @analyses = Analysis.order(created_at: :desc).page(1)
     end
+  end
 
-    after(:all) { Analysis.destroy_all }
+  shared_examples '画面共通テスト' do |expected: {}|
+    it_behaves_like 'ヘッダーが表示されていること'
+    it_behaves_like '入力フォームが表示されていること'
+    it_behaves_like '表示件数情報が表示されていること',
+                    total: expected[:total] || per_page,
+                    from: expected[:from] || 1,
+                    to: expected[:to] || per_page
+    it_behaves_like 'テーブルが表示されていること',
+                    rows: expected[:rows] || per_page
   end
 
   shared_examples '入力フォームが表示されていること' do
     form_xpath = '//form[@id="new_analysis"]'
 
-    %w[ from to batch_size ].each do |param|
+    %w[from to pair batch_size].each do |param|
       input_xpath = "#{form_xpath}/div[@class='form-group']"
 
       it "analysis_#{param}を含む<label>タグがあること" do
-        expect(@html).to have_selector("#{input_xpath}/label[for='analysis_#{param}']")
+        label = @html.xpath("#{input_xpath}/label[@for='analysis_#{param}']")
+        is_asserted_by { label.present? }
       end
 
-      it "analysis_#{param}を含む<input>タグがあること" do
-        expect(@html).to have_selector("#{input_xpath}/input[id='analysis_#{param}']")
+      it "analysis_#{param}を含む<input>タグがあること", unless: param == 'pair' do
+        input = @html.xpath("#{input_xpath}/input[@id='analysis_#{param}']")
+        is_asserted_by { input.present? }
+      end
+
+      it "analysis_#{param}を含む<select>タグがあること", if: param == 'pair' do
+        select = @html.xpath("#{input_xpath}/select[@id='analysis_#{param}']")
+        is_asserted_by { select.present? }
+      end
+
+      Analysis::PAIRS.each do |pair|
+        it "#{pair}を選択できること", if: param == 'pair' do
+          select_xpath = "#{input_xpath}/select[@id='analysis_#{param}']"
+          option = @html.xpath("#{select_xpath}/option[@value='#{pair}']")
+          is_asserted_by { option.present? }
+        end
+      end
+
+      it 'デフォルトでUSDJPYが選択されていること', if: param == 'pair' do
+        select_xpath = "#{input_xpath}/select[@id='analysis_#{param}']"
+        default_option = @html.xpath("#{select_xpath}/option[@value='USDJPY']")
+        is_asserted_by { default_option.attribute('selected').value == 'selected' }
       end
     end
 
-    %w[ submit reset ].each do |type|
+    %w[submit reset].each do |type|
       it "typeが#{type}のボタンがあること" do
-        expect(@html).to have_selector("#{form_xpath}/input[type='#{type}']")
+        button = @html.xpath("#{form_xpath}/input[@type='#{type}']")
+        is_asserted_by { button.present? }
       end
     end
   end
 
-  shared_examples 'ジョブ実行履歴が表示されていること' do |expected_size: per_page, total: 0, from: 1, to: 1|
-    base_xpath = '/html/body/div[@id="main-content"]/div[@class="row center-block"]/div[@class="col-lg-8"]'
-
-    it 'タイトルが表示されていること' do
-      expect(@html).to have_selector("#{base_xpath}/h3", :text => 'ジョブ実行履歴')
+  shared_examples 'テーブルが表示されていること' do |rows: 0|
+    before(:each) do
+      @table = @html.xpath("#{table_panel_xpath}/table[@class='table table-hover']")
     end
 
-    it '件数情報が表示されていること' do
-      expect(@html).to have_selector("#{base_xpath}/h4", :text => "#{total}件中#{from}〜#{to}件を表示")
+    it '6列のテーブルが表示されていること' do
+      is_asserted_by { @table.xpath('//thead/th').size == 6 }
     end
 
-    paging_xpath = "#{base_xpath}/nav/ul[@class='pagination']"
-
-    it '先頭のページへのボタンが表示されていないこと' do
-      xpath = "#{paging_xpath}/li[@class='pagination']/span[@class='first']/a"
-      expect(@html).not_to have_selector(xpath, :text => I18n.t('views.list.pagination.first'))
-    end
-
-    it '前のページへのボタンが表示されていないこと' do
-      xpath = "#{paging_xpath}/li[@class='pagination']/span[@class='prev']/a"
-      expect(@html).not_to have_selector(xpath, :text => I18n.t('views.list.pagination.previous'))
-    end
-
-    it '1ページ目が表示されていること' do
-      xpath = "#{paging_xpath}/li[@class='page-item active']"
-      expect(@html).to have_selector(xpath, :text => 1)
-    end
-
-    it '2ページ目が表示されていること' do
-      xpath = "#{paging_xpath}/li[@class='page-item']/a[href='/analyses?page=2']"
-      expect(@html).to have_selector(xpath, :text => 2)
-    end
-
-    it '次のページへのボタンが表示されていること' do
-      xpath = "#{paging_xpath}/li[@class='page-item']/span[@class='next']/a[href='/analyses?page=2']"
-      expect(@html).to have_selector(xpath, :text => I18n.t('views.pagination.next'))
-    end
-
-    it '最後のページへのボタンが表示されていること' do
-      xpath = "#{paging_xpath}/li[@class='page-item']/span[@class='last']/a"
-      expect(@html).to have_selector(xpath, :text => I18n.t('views.pagination.last'))
-    end
-
-    %w[ 実行開始日時 期間 バッチサイズ 状態 ].each do |header|
-      it "ヘッダー(#{header})があること" do
-        xpath = "#{base_xpath}/table[@class='table table-hover']/thead/th"
-        expect(@html).to have_selector(xpath, :text => header)
+    %w[実行開始日時 期間 ペア バッチサイズ 状態].each_with_index do |text, i|
+      it "#{i + 1}列目のヘッダーが#{text}であること" do
+        is_asserted_by { @table.xpath('//thead/th')[i].text == text }
       end
     end
 
     it '再実行ボタンが配置されている列があること' do
-        xpath = "#{base_xpath}/table[@class='table table-hover']/thead/th[@class='rebuild']"
-        expect(@html).to have_selector(xpath, :text => '')
+      xpath = [
+        table_panel_xpath,
+        'table[@class="table table-hover"]',
+        'thead',
+        'th[@class="rebuild"]',
+      ].join('/')
+      header_rebuild = @table.xpath(xpath)
+      is_asserted_by { header_rebuild.present? }
     end
 
-    it 'データの数が正しいこと' do
-      xpath = "#{base_xpath}/table[@class='table table-hover']/tbody/tr"
-      expect(@html).to have_xpath(xpath, :count => expected_size)
+    it 'ジョブの数が正しいこと' do
+      is_asserted_by { @table.xpath('//tbody/tr').size == rows }
+    end
+  end
+
+  shared_examples 'ジョブの状態が正しいこと' do |state|
+    before(:each) do
+      @rows =
+        @html.xpath("#{table_panel_xpath}/table[@class='table table-hover']/tbody/tr")
     end
 
-    it '背景色が正しいこと', :if => expected_size > 0 do
-      matched_data = @html.gsub("\n", '').match(/<td\s*class='(?<color>(warning|success))'\s*>(?<state>(実行中|完了))<\/td>/)
-      expected = {'warning' => '実行中', 'success' => '完了'}
-      is_asserted_by { expected[matched_data[:color]] == matched_data[:state] }
+    it 'ペアが正しいこと' do
+      @rows.each_with_index do |row, i|
+        displayed_pair = row.children.search('td')[2].text.strip
+        is_asserted_by { displayed_pair == @analyses[i].pair }
+      end
+    end
+
+    it '状態が正しいこと' do
+      @rows.each do |row|
+        is_asserted_by { row.xpath('//td')[4].text.strip == state }
+      end
     end
   end
 
@@ -109,19 +131,45 @@ describe 'analyses/manage', :type => :view do
   end
 
   before(:each) do
-    render :template => 'analyses/manage', :layout => 'layouts/application'
-    @html ||= response
+    render template: 'analyses/manage', layout: 'layouts/application'
+    @html ||= Nokogiri.parse(response)
   end
 
-  describe '<html><body>' do
-    include_context '分析ジョブを登録する', 10
-    it_behaves_like 'ヘッダーが表示されていること'
-    it_behaves_like '入力フォームが表示されていること'
-    it_behaves_like 'ジョブ実行履歴が表示されていること', {
-                      :expected_size => 1,
-                      :total => 10,
-                      :from => 1,
-                      :to => 1,
-                    }
+  context '実行中の場合' do
+    include_context 'トランザクション作成'
+    include_context '分析ジョブを登録する'
+    include_context 'HTML初期化'
+    it_behaves_like '画面共通テスト'
+    it_behaves_like 'ページングボタンが表示されていないこと'
+    it_behaves_like 'ジョブの状態が正しいこと', '実行中'
+  end
+
+  context '完了しているの場合' do
+    attribute = default_attribute.merge(state: 'completed')
+    include_context 'トランザクション作成'
+    include_context '分析ジョブを登録する', attribute: attribute
+    include_context 'HTML初期化'
+    it_behaves_like '画面共通テスト'
+    it_behaves_like 'ページングボタンが表示されていないこと'
+    it_behaves_like 'ジョブの状態が正しいこと', '完了'
+  end
+
+  context 'エラーの場合' do
+    attribute = default_attribute.merge(state: 'error')
+    include_context 'トランザクション作成'
+    include_context '分析ジョブを登録する', attribute: attribute
+    include_context 'HTML初期化'
+    it_behaves_like '画面共通テスト'
+    it_behaves_like 'ページングボタンが表示されていないこと'
+    it_behaves_like 'ジョブの状態が正しいこと', 'エラー'
+  end
+
+  context "分析ジョブ情報が#{per_page * (Kaminari.config.window + 2)}件の場合" do
+    total = per_page * (Kaminari.config.window + 2)
+    include_context 'トランザクション作成'
+    include_context '分析ジョブを登録する', total: total
+    include_context 'HTML初期化'
+    it_behaves_like '画面共通テスト', expected: {total: total}
+    it_behaves_like 'ページングボタンが表示されていること', model: 'analysis'
   end
 end